feat: search, comments, and refactored hook for business logic
