+++ src/app/api/advocates/route.ts
@@ -1,4 +1,4 @@
-import { ilike, or, and } from "drizzle-orm";
+import { ilike, or, sql } from "drizzle-orm";
 import db from "../../../db";
 import { advocates } from "../../../db/schema";
 
@@ -6,17 +6,21 @@ export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
   const filter = searchParams.get("q");
 
+  // TODO: paginate, sort, and better search... ilikes prolly not the best eh?
   const data = await db
     .select()
     .from(advocates)
+    // Annoying because in the docs they dont have an issue here: https://orm.drizzle.team/docs/guides/conditional-filters-in-query
     .where(
-      and(
-        or(
-          ilike(advocates.firstName, `%${filter}%`),
-          ilike(advocates.lastName, `%${filter}%`),
-          ilike(advocates.city, `%${filter}%`),
-          ilike(advocates.degree, `%${filter}%`)
-        )
+      or(
+        ilike(advocates.firstName, `%${filter}%`),
+        ilike(advocates.lastName, `%${filter}%`),
+        ilike(advocates.city, `%${filter}%`),
+        ilike(advocates.degree, `%${filter}%`),
+        // Get out of jail here casting as strings...
+        sql`${advocates.specialties}::text ILIKE ${`%${filter}%`}`,
+        sql`${advocates.yearsOfExperience}::text ILIKE '%' || ${filter} || '%'`,
+        sql`${advocates.phoneNumber}::text ILIKE '%' || ${filter} || '%'`
       )
     );
